investor instructions almost fixedgit add .git add .git add .!
