+++ app/src/main/java/com/example/myapplication/Investor_Instructions.java
@@ -1,25 +1,31 @@
 package com.example.myapplication;
 
+import Objects.Manager;
 import Objects.Share;
 import Objects.ShareAdapter;
+import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.View;
-import android.widget.ArrayAdapter;
 import android.widget.Button;
 import android.widget.ListView;
-import android.widget.TableLayout;
-import android.widget.TextView;
 
-import com.example.myapplication.Business_Logic.Investor_Logic;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.Query;
+import com.google.firebase.database.ValueEventListener;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 
 public class Investor_Instructions extends AppCompatActivity {
     String investor_id;
@@ -42,17 +48,15 @@ public class Investor_Instructions extends AppCompatActivity {
         tableLayout=findViewById(R.id.company_shares_table_investor_instructions);
         investor_id = getIntent().getStringExtra("user_id");
         investor_shares = new ArrayList<>();
-        ShareAdapter shareAdapter =new ShareAdapter(context,investor_shares);
+        final ShareAdapter shareAdapter =new ShareAdapter(context,investor_shares);
         tableLayout.setAdapter(shareAdapter);
-        IL= new Investor_Logic(investor_id, this ,shareAdapter);
+        IL= new Investor_Logic(investor_id,shareAdapter);
+        IL.get_symbols();
 
         SRL.setOnRefreshListener(new SwipeRefreshLayout.OnRefreshListener() {
             @Override
             public void onRefresh() {
-                IL.get_symbols();
-
-
-
+                shareAdapter.clear();
 
             }
 
@@ -87,7 +91,104 @@ public class Investor_Instructions extends AppCompatActivity {
     }
 
 
+    public static class Investor_Logic {
+
+        private final static String TAG="INVESTOR_LOGIC";
+
+        private String investor_id;
+        private HashMap<String, String> symbol_id;
+        private ShareAdapter shareAdapter;
+
+        public Investor_Logic(String investor_id,  ShareAdapter shareAdapter) {
+           this.investor_id =investor_id;
+
+           this.shareAdapter = shareAdapter;
 
 
 
+        }
+
+
+
+        public HashMap<String, String> getSymbol_id() {
+            return symbol_id;
+        }
+
+
+        public void get_symbols(){
+           symbol_id= new HashMap<>();
+            FirebaseDatabase db = FirebaseDatabase.getInstance();
+            Query q = db.getReference("Managers");
+
+            q.addListenerForSingleValueEvent(new ValueEventListener() {
+                @Override
+                public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
+                    for (DataSnapshot d : dataSnapshot.getChildren()){
+                        for (DataSnapshot manager :dataSnapshot.getChildren()){
+                            String manager_id = manager.getKey();
+                            String company_symbol = manager.getValue(Manager.class).getCompany_symbol();
+                            symbol_id.put(company_symbol,manager_id);
+
+                        }
+
+                    }
+                   retrieve_investor_data(symbol_id);
+
+
+                }
+
+                @Override
+                public void onCancelled(@NonNull DatabaseError databaseError) {
+                    Log.d(TAG, "******************************can't get the investors?***********************");
+
+                }
+            });
+
+
+        }
+
+
+
+
+        public void retrieve_investor_data(HashMap<String, String> symbol_id) {
+            Log.d(TAG, "IN RETRIEVE INVESTOR DATA");
+
+
+
+
+
+            for (final String sym: symbol_id.keySet()){
+                final String manager_val=symbol_id.get(sym);
+                final Query shares =FirebaseDatabase.getInstance().getReference("Shares").child(investor_id).child(sym);
+                shares.addValueEventListener(new ValueEventListener() {
+                    @Override
+                    public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
+                        Share share = dataSnapshot.getValue(Share.class);
+                        if (share==null){
+                            DatabaseReference r = FirebaseDatabase.getInstance().getReference("Shares").child(investor_id).child(sym);
+                            r.setValue(new Share(investor_id,sym, manager_val ));
+                            shareAdapter.add(new Share(investor_id,sym, manager_val ));
+
+                        }else shareAdapter.add(share);
+
+                    }
+
+                    @Override
+                    public void onCancelled(@NonNull DatabaseError databaseError) {
+
+                    }
+
+                });
+
+
+
+             }
+
+
+
+
+
+
+        }
+    }
 }
