+++ app/src/main/res/layout/activity_investor__instructions.xml
@@ -37,7 +37,8 @@
          android:layout_height="wrap_content" />
     <LinearLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
+        android:layout_height="wrap_content"
+        app:layout_constraintTop_toBottomOf="@id/investor_instructions">
 
    <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
         android:id="@+id/swiper_investor_instructions"
