+++ app/src/main/res/layout/res_share.xml
@@ -1,6 +1,9 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
-    android:layout_height="match_parent" >
+    android:layout_height="match_parent"
+    android:layoutDirection="inherit"
+    >
+
     <TextView
         android:id="@+id/compsymb"
         android:layout_width="wrap_content"
